+++ go/pkg/basecamp/checkins.go
@@ -660,7 +660,7 @@ func questionnaireFromGenerated(gq generated.Questionnaire) Questionnaire {
 
 	if gq.Creator.Id != 0 || gq.Creator.Name != "" {
 		q.Creator = &Person{
-			ID:           gq.Creator.Id,
+			ID:           int64(gq.Creator.Id),
 			Name:         gq.Creator.Name,
 			EmailAddress: gq.Creator.EmailAddress,
 			AvatarURL:    gq.Creator.AvatarUrl,
@@ -744,7 +744,7 @@ func questionFromGenerated(gq generated.Question) Question {
 
 	if gq.Creator.Id != 0 || gq.Creator.Name != "" {
 		q.Creator = &Person{
-			ID:           gq.Creator.Id,
+			ID:           int64(gq.Creator.Id),
 			Name:         gq.Creator.Name,
 			EmailAddress: gq.Creator.EmailAddress,
 			AvatarURL:    gq.Creator.AvatarUrl,
@@ -804,7 +804,7 @@ func questionAnswerFromGenerated(ga generated.QuestionAnswer) QuestionAnswer {
 
 	if ga.Creator.Id != 0 || ga.Creator.Name != "" {
 		a.Creator = &Person{
-			ID:           ga.Creator.Id,
+			ID:           int64(ga.Creator.Id),
 			Name:         ga.Creator.Name,
 			EmailAddress: ga.Creator.EmailAddress,
 			AvatarURL:    ga.Creator.AvatarUrl,
